% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_this}
\alias{download_this}
\title{Download button}
\usage{
download_this(
  .data,
  output_name,
  output_extension = c(".csv", ".xlsx"),
  button_label,
  button_type = c("default", "primary", "success", "info", "warning", "danger"),
  ...
)
}
\arguments{
\item{.data}{A data frame to write to disk.}

\item{output_name}{Name of of the output file.}

\item{output_extension}{Extension of the output file. Currently, only  \code{.csv} and  \code{.xlsx} are supported.}

\item{button_label}{character (HTML), button label}

\item{button_type}{character, one of the standard Bootstrap types}

\item{...}{attributes (named arguments) and children (unnamed arguments)
of the button, passed to \code{htmltools::\link[htmltools]{tag}}}
}
\value{
\code{htmltools::\link[htmltools]{tag}}, \code{<button/>}
}
\description{
Wrapper around \code{bsplus::\link[bsplus]{bs_button}} to provide a download button for HTML outputs in RMarkdown
}
\examples{

mtcars \%>\%
  download_this(
    output_name = "mtcars data set",
    output_extension = ".csv",
    button_label = "Download data",
    button_type = "warning",
  )
}
