% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadthis-package.R
\docType{package}
\name{downloadthis-package}
\alias{downloadthis}
\alias{downloadthis-package}
\title{downloadthis: Implement Download Buttons in 'rmarkdown'}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Implement download buttons in HTML output from 'rmarkdown' without the need for 'runtime:shiny'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/fmmattioni/downloadthis}
  \item Report bugs at \url{https://github.com/fmmattioni/downloadthis/issues}
}

}
\author{
\strong{Maintainer}: Felipe Mattioni Maturana \email{felipe.mattioni@med.uni-tuebingen.de} (\href{https://orcid.org/0000-0002-4221-6104}{ORCID})

Other contributors:
\itemize{
  \item John Coene \email{jcoenep@gmail.com} [contributor]
}

}
\keyword{internal}
